data_analyst_agent/streamlit/requirements.txt
streamlit
pandas
matplotlib
llama-index
llama-index-llms-gemini
llama-index-experimental
google-generativeai